Added sudoku class and refactored previous functions into sudoku_utils.py
